[WIP] TIP-987: PublishedProductInterface does not extend ProductInterface anymore
